// src/lib/translator-sdk/models/translate-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GPT-Translator
 * The GPT translator API description
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface TranslateDto
 */
export interface TranslateDto {

    /**
     * The text to translate
     *
     * @type {string}
     * @memberof TranslateDto
     */
    inputMessage: string;

    /**
     * The language to translate from
     *
     * @type {string}
     * @memberof TranslateDto
     */
    inputLanguage: string;

    /**
     * The language to translate to
     *
     * @type {string}
     * @memberof TranslateDto
     */
    outputLanguage: string;
}